Refactor start and end minute rule decorator to just be a function [#65]
