+++ src/assistants/buyback.ts
@@ -1,9 +1,9 @@
+import betweenMinutes from "../engine/rules/betweenMinutes";
 import configurable from "../engine/rules/configurable";
 import { EffectConfig } from "../effectConfigManager";
 import Fact from "../engine/Fact";
 import inGame from "../engine/rules/inGame";
 import Rule from "../engine/Rule";
-import RuleDecoratorStartAndEndMinute from "../engine/rules/RuleDecoratorStartAndEndMinute";
 import rules from "../rules";
 import topicManager from "../engine/topicManager";
 import topics from "../topics";
@@ -20,7 +20,7 @@ const hasGoldForBuybackTopic = topicManager.createTopic<boolean>(
 );
 
 export default [
-    new RuleDecoratorStartAndEndMinute(
+    betweenMinutes(
         30,
         undefined,
         new Rule({
