+++ src/assistants/runePower.ts
@@ -1,9 +1,9 @@
+import betweenMinutes from "../engine/rules/betweenMinutes";
 import configurable from "../engine/rules/configurable";
 import { EffectConfig } from "../effectConfigManager";
 import Fact from "../engine/Fact";
 import inGame from "../engine/rules/inGame";
 import Rule from "../engine/Rule";
-import RuleDecoratorStartAndEndMinute from "../engine/rules/RuleDecoratorStartAndEndMinute";
 import rules from "../rules";
 import topicManager from "../engine/topicManager";
 import topics from "../topics";
@@ -17,7 +17,7 @@ export const defaultConfig = EffectConfig.PRIVATE;
 export const assistantDescription =
     "Reminds you of power rune spawn every 2:00 after 6:00";
 
-export default new RuleDecoratorStartAndEndMinute(
+export default betweenMinutes(
     6,
     undefined,
     inGame(
